// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferObjectMailFolderCountViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */

export interface TransferObjectMailFolderCountViewDTO {
  /** 数量。 */
  count?: number;
  /** 邮件文件夹。 */
  mailFolder?: string;
  /** 邮件文件夹标签。 */
  mailFolderLabel?: string;
}